requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
arrow==1.3.0
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
binaryornot==0.4.4
bleach==6.2.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cookiecutter==2.6.0
cron-descriptor==1.4.5
Django==5.2.5
django-browser-reload==1.18.0
django-celery-beat==2.8.1
django-environ==0.12.0
django-redisboard==9.0.0
django-tailwind==4.2.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.31.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
honcho==2.0.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kombu==5.5.4
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==25.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.3
pyparsing==3.2.3
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.5
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.2
text-unidecode==1.3
tqdm==4.67.1
types-python-dateutil==2.9.0.20250809
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
